+++ modules/apps/headless/headless-data-masking/headless-data-masking-api/src/main/java/com/liferay/headless/data/masking/engine/DataMaskingEngine.java
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.headless.data.masking.service.v1_0;
+package com.liferay.headless.data.masking.engine;
 
 import java.util.List;
 
 /**
  * @author Jose Luis Navarro
  */
-public interface DataMaskingService {
+public interface DataMaskingEngine {
 
 	public void evictPattern(String regex);
