+++ modules/apps/headless/headless-data-masking/headless-data-masking-impl/src/test/java/com/liferay/headless/data/masking/internal/engine/DataMaskTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.headless.data.masking.internal.masking;
+package com.liferay.headless.data.masking.internal.engine;
 
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
