+++ modules/apps/mcp/mcp-server-rest-impl/src/main/java/com/liferay/mcp/server/rest/internal/servlet/MCPServerServlet.java
@@ -7,7 +7,7 @@ package com.liferay.mcp.server.rest.internal.servlet;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import com.liferay.headless.data.masking.service.v1_0.DataMaskingService;
+import com.liferay.headless.data.masking.engine.DataMaskingEngine;
 import com.liferay.mcp.server.rest.dto.v1_0.Tool;
 import com.liferay.mcp.server.rest.internal.configuration.MCPServerConfiguration;
 import com.liferay.mcp.server.rest.internal.constants.MCPServerConstants;
@@ -265,7 +265,7 @@ public class MCPServerServlet extends HttpServlet {
 				List<String> maskExternalReferenceCodes = _getDataMaskERCs(
 					companyId, profileObjectEntryId);
 
-				content = _dataMaskingService.redact(
+				content = _dataMaskingEngine.redact(
 					companyId, maskExternalReferenceCodes, content);
 			}
 
@@ -515,7 +515,7 @@ public class MCPServerServlet extends HttpServlet {
 	private ConfigurationProvider _configurationProvider;
 
 	@Reference
-	private DataMaskingService _dataMaskingService;
+	private DataMaskingEngine _dataMaskingEngine;
 
 	@Reference(
 		target = "(filter.factory.key=" + ObjectDefinitionConstants.STORAGE_TYPE_DEFAULT + ")"
